# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VoidMind API
  description: |
    Zero-Knowledge AI Gateway — Think. Respond. Vanish.
    Stateless AI inference with RAM-only sessions.
  version: 1.0.0
  contact:
    name: Nuyvo LLC
    url: https://github.com/nuyvo/voidmind
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your-domain.com/api/v1
    description: Production
  - url: http://127.0.0.1:3000/api/v1
    description: Local Development

paths:
  /health:
    get:
      summary: Health Check
      description: Public health endpoint. No authentication required.
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  service: { type: string, example: voidmind }
                  version: { type: string, example: 1.0.0 }
                  uptime: { type: number }
                  timestamp: { type: string, format: date-time }

  /chat/completions:
    post:
      summary: Chat Completions
      description: Main inference endpoint. OpenAI-compatible format.
      tags: [Chat]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: Existing session ID or omit to create new
                model:
                  type: string
                  default: qwen2.5:3b
                messages:
                  type: array
                  maxItems: 20
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant]
                      content:
                        type: string
                        maxLength: 4096
                temperature:
                  type: number
                  minimum: 0
                  maximum: 2
                  default: 0.7
                max_tokens:
                  type: integer
                  minimum: 1
                  maximum: 8192
                  default: 512
                stream:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Chat completion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  object: { type: string, example: chat.completion }
                  created: { type: integer }
                  model: { type: string }
                  session_id: { type: string }
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        index: { type: integer }
                        message:
                          type: object
                          properties:
                            role: { type: string }
                            content: { type: string }
                        finish_reason: { type: string }
                  usage:
                    type: object
                    properties:
                      prompt_tokens: { type: integer }
                      completion_tokens: { type: integer }
                      total_tokens: { type: integer }
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream (when stream=true)
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded

  /session/end:
    post:
      summary: End Session
      description: Explicitly wipe a session from RAM.
      tags: [Sessions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Session ended

  /models:
    get:
      summary: List Models
      description: List available Ollama models.
      tags: [Models]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of models

  /usage:
    get:
      summary: Key Usage
      description: Token usage for the authenticated API key.
      tags: [Usage]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage statistics

  /admin/auth/login:
    post:
      summary: Admin Login
      description: Authenticate admin and receive JWT tokens.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Tokens issued

  /admin/keys:
    get:
      summary: List API Keys
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API keys
    post:
      summary: Create API Key
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                monthly_limit: { type: integer }
                daily_limit: { type: integer }
                requests_per_minute: { type: integer }
      responses:
        '201':
          description: Key created

  /admin/sessions:
    get:
      summary: Active Sessions
      description: Session counts and metadata (no content).
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session statistics
    delete:
      summary: Force Wipe All Sessions
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All sessions wiped

  /admin/health:
    get:
      summary: Admin Health
      description: Detailed system health metrics.
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Health details

  /admin/compliance:
    get:
      summary: Compliance Report
      description: Live compliance and audit report.
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Compliance data

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT